+++ src/reanimated2/animation/decay/index.ts
@@ -1,0 +1,3 @@
+export { withDecay } from './decay';
+export type { WithDecayConfig } from './decay';
+export type { DecayAnimation } from './utils';
